+++ moz_security/tests/hindsight/run/analysis/hh_monitor_tigerblood.cfg
@@ -1,0 +1,9 @@
+filename = "moz_security_heavy_hitters_tigerblood.lua"
+message_matcher = "Logger == 'input.hh_tigerblood'"
+ticker_interval = 0
+
+id_field = "Fields[id]"
+hh_items = 10
+
+tigerblood = {}
+violation_type = "fxa:heavy-hitter-ip"
