+++ moz_security/tests/hindsight/run/input/generate_hh_tigerblood.cfg
@@ -1,0 +1,1 @@
+filename = "generate_hh_tigerblood.lua"
