+++ moz_security/tests/hindsight/run/output/hh_tigerblood_verification.cfg
@@ -1,0 +1,2 @@
+filename = "hh_tigerblood_verification.lua"
+message_matcher = "Logger == 'analysis.hh_monitor_tigerblood'"
